refactor: Tracing
The `lib.rs` file accumulated structs and functionality from different areas.
Split the file into several smaller files.
