+++ core/o11y/src/log_config.rs
@@ -1,3 +1,4 @@
+use crate::OpenTelemetryLevel;
 use serde::{Deserialize, Serialize};
 use std::path::Path;
 use std::{fs::File, io::Write};
@@ -11,7 +12,7 @@ pub struct LogConfig {
     /// Some("module") enables debug logging for "module".
     pub verbose_module: Option<String>,
     /// Verbosity level of collected traces.
-    pub opentelemetry_level: Option<crate::OpenTelemetryLevel>,
+    pub opentelemetry_level: Option<OpenTelemetryLevel>,
 }
 
 impl LogConfig {
